+++ src/DialogConfig.h
@@ -1,3 +1,9 @@
+//======================================================================*/
+//   Create Date: 2022/11/15
+//   ClassName  : Daqmain
+//   Author     : mhjin julu1@naver.com
+//   Class Def  : configration settion을 위한 dialog window
+//======================================================================*/
 #ifndef DIALOGCONFIG_H
 #define DIALOGCONFIG_H
 
@@ -10,7 +16,11 @@ class DialogConfig;
 class DialogConfig : public QDialog
 {
     Q_OBJECT
-
+    public slots:
+    void setScriptDir();
+    void setRvizDir();
+    void setRecordDir();
+    void accept();
 public:
     explicit DialogConfig(QWidget *parent = nullptr);
     ~DialogConfig();